% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ref_timelines.R
\name{add.timeline}
\alias{add.timeline}
\title{Plot a timeline segment}
\usage{
add.timeline(x0, x1, y, notch.height = 0.5, padd.x = 0, ...)
}
\arguments{
\item{x0, x1}{coordinates of start and end of the segment.}

\item{y}{y coordinate of the segment.}

\item{notch.height}{length of the edge notches.}

\item{padd.x}{padding ; the segment is shortened by \code{padd.x/2} on each end}

\item{...}{arguments passed on to \code{\link[graphics]{segments}}}
}
\description{
Adds horizontal segments with notches at each end to the current plot.
}
\examples{
\donttest{
plot(1:5, type='n', )
add.timeline(x0 = c(1,3), x1 = c(4,5), y = c(2,3), lwd=2)
}

}
