% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae.R
\name{ae}
\alias{ae}
\title{Age Estimate}
\usage{
ae(
  samp,
  refdata,
  ref.time_series = NULL,
  cor.method = "spearman",
  nb.cores = 2,
  bootstrap.n = 30,
  bootstrap.set_size = NULL,
  prior = NULL,
  prior.params = NULL,
  verbose = T
)
}
\arguments{
\item{samp}{sample matrix, genes as rows, individuals as columns}

\item{refdata}{a reference object, as returned by \code{\link{make_ref}}, or a reference time series matrix in same format as \code{samp}}

\item{ref.time_series}{ignored if a \code{ref} object is given to \code{refdata}, else the time values of the reference data specified)}

\item{cor.method}{correlation method, one of "spearman" (default) or "pearson".}

\item{nb.cores}{number of cores for parallelism, defaults to 2.}

\item{bootstrap.n}{number of bootstraps. Defaults to 30, should be >5.}

\item{bootstrap.set_size}{random gene set size for the bootstrap, defaults to n/3 (with n, the size of the sample-reference gene set \emph{overlap}).}

\item{prior}{Approximate time values for each sample, in the time unit of the given reference. Values are recycled if less than the number of samples}

\item{prior.params}{Standard deviation of the prior scoring distribution. \emph{Setting this value too low can cause a significant bias in the age estimation.}}

\item{verbose}{if TRUE (default), displays progression messages.}
}
\value{
an \code{ae} object, with the age estimates, the correlation matrix between sample and reference,
the reference time series, and the bootstrap age estimates and correlation matrices.
There are \code{plot}, \code{print} and \code{summary} methods for this object.
}
\description{
\code{ae} estimates the developmental age of input samples based on
correlation with the given reference data.
}
\details{
Confidence intervals for each estimate are computed from bootstrapping on genes and computing
the Median Absolute Deviation of the bootstrap age estimates to the global estimate.

A prior can be given to help with the estimate, in which case the peaks
of the correlation profiles will be scored according to a gaussian
of the specified parameters.
}
\examples{
\donttest{
requireNamespace('wormRef', quietly = TRUE)

# get some samples to stage
samp <- wormRef::Cel_larval$g[,13:15]

# load interpolated reference
r_larv <- prepare_refdata(ref = 'Cel_larval', datapkg = 'wormRef' , n.inter = 200)

# perform age estimate
ae_test <- ae(samp, r_larv)

# check output
summary(ae_test)
plot(ae_test) # plot all sample estimates
plot_cor(ae_test) # plot individual correlation profiles of samples

# get results
ae_test$age.estimates
}
}
