% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cel_larval-data.R
\docType{data}
\name{Cel_larval}
\alias{Cel_larval}
\title{C. elegans Larval development reference data time series}
\format{a list with \code{X} and \code{X.25} being the gene expression matrices, \code{time.series} and \code{time.series.25} being the timepoints (in hours post-hatching). \code{est.time.series} is the re-estimated age for timepoints of both time series together (scaled on the 20C time).}
\source{
\href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE49043}{GEO dataset}
}
\usage{
data(Cel_larval)
}
\description{
Data from a paper published by Oudenaarden et al.
with C elegans gene expression levels (log+1 from rpkm).

This dataset is a reference time series for C. elegans development at both 20C and 25C
from hatching to late L4 (48h at 20C)
}
\examples{
data(Cel_larval)
times <- Cel_larval$time.series
gene1 <- Cel_larval$X[1,]
\donttest{plot(times, gene1, type='l')}
}
\references{
Oudenaarden et al. (2013) Nat Genet. 2013 Nov; 45(11): 1337–1344.
(\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3812263/}{PubMed})
}
\keyword{datasets}
