% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cel_YA_adult1-data.R
\docType{data}
\name{Cel_YA_adult1}
\alias{Cel_YA_adult1}
\title{C. elegans Young Adult/Adult reference data time series (1)}
\format{a list with \code{X} being the gene expression matrix and \code{time.series} being the timepoints (in hours post-hatching)}
\source{
\href{https://www.ebi.ac.uk/arrayexpress/experiments/E-MTAB-7574/}{ArrayExpress dataset}
}
\usage{
data(Cel_YA_adult1)
}
\description{
Data from the ArrayExpress submitted by Mark G. Sterken in 2014
(recently published by Van Sluijs et al. (2019)), with C elegans gene expression
levels (log+1 from intensities).
}
\details{
Only samples of Standard N2 strain were kept here.

Time is in hours from hatching, scaled on the Oudenaarden 20C time series, 
re-estimated on the Reinke and Oudenaarden references.
}
\examples{
data(Cel_YA_adult1)
times <- Cel_YA_adult1$time.series
gene1 <- Cel_YA_adult1$X[700,]
\donttest{plot(times, gene1, type='b')}
}
\references{
Van Sluijs, et al.  (2019) BioRxiv : 579151
(\href{https://www.biorxiv.org/content/10.1101/579151v3}{bioRxiv})
}
\keyword{datasets}
