% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coefficients}
\alias{coefficients}
\title{Coefficients for 4 predictors of immunotherapy benefit}
\format{
A list of 4 named numeric vectors.
\describe{
 \item{coefficients}{Numeric values giving the weight of each gene within each signature}

}
}
\source{

}
\usage{
coefficients
}
\description{
The coefficents required to generate the 4 probability of response scores.
}
\keyword{datasets}
