% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeRaptor.R
\name{compute_risk_score}
\alias{compute_risk_score}
\title{Get a gene expression matrix with genes as colnames, and samples as rownames.}
\usage{
compute_risk_score(expr, coefs, predictor_name, constant)
}
\arguments{
\item{expr}{a data frame of gene expression data with genes as column names and samples as rownames.
Ideally expression will be in log2(TPM+1) format. The data frame should also include columns for the MSKCC prognostic score
if the data is from renal cancer patients (column name = "mskcc").}

\item{coefs}{a named numeric vector of gene weights.}

\item{predictor_name}{the name of the probability of response score estimator, for example "RAPTOR"}

\item{constant}{a constant to add to the scores. We will use the minimum value of the sum of the weighted gene expressions
in the training data multiplied by -1 if less than 0.}
}
\value{
dataframe with sample as rowname and column as predictor score
}
\description{
This function will score the likelihood of patient response to immune checkpoint blockade based on 4 different pipelines.
1. RAPTOR - based on a set of immune genes
2. COXIS - based on a 25-gene set defined by Bonavtia et al., in review
3. TIS - based on an 18 gene signature from Ayers et al., JCI, 2017
4. Random - a random set of genes to which identical steps were applied as for the other pipelines
}
\details{
If more than 50% of variables are missing then the exercise should be halted
}
