% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{constant_list}
\alias{constant_list}
\title{Constants for 4 predictors of immunotherapy benefit}
\format{
A list of 4 named numeric vectors.
\describe{
 \item{constant_list}{Numeric values giving the weight of each gene within each signature}

}
}
\source{

}
\usage{
constant_list
}
\description{
The constants required to generate the 4 probability of response scores.
}
\keyword{datasets}
