% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_CV.R
\name{ef}
\alias{ef}
\title{CV error function}
\usage{
ef(xt, xe)
}
\arguments{
\item{xt, xe}{vectors of validation data and predictions respectively.}
}
\value{
The sum squared prediction error.
}
\description{
Computes sum((xt-xe)^2)
}
