% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae.R
\name{estimate.worm_age}
\alias{estimate.worm_age}
\title{Age Estimate - \strong{DEPRECATED}}
\usage{
estimate.worm_age(samp, refdata, ref.time_series, ...)
}
\arguments{
\item{samp}{passed on to ae}

\item{refdata}{passed on to ae}

\item{ref.time_series}{passed on to ae}

\item{...}{passed on to ae}
}
\description{
\strong{DEPRECATED}. Please use \code{\link{ae}} instead.
}
