% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_ids-data.R
\docType{data}
\name{gene_ids}
\alias{gene_ids}
\title{Gene ID table}
\format{a dataframe with corresponding WBGene IDs (\code{WormBase.Gene.ID}), public names (\code{Public.Name}) and sequence names (\code{Sequence.Name}).}
\source{
\href{https://www.wormbase.org/}{WormBase}
}
\usage{
data(gene_ids)
}
\description{
A dataframe with WBGene IDs, common public name and Sequence name.
This table may be especially useful to convert a dataset's row ids to
WBGene IDs, as needed to find the overlapping geneset with the reference
and perform the age estimation.
}
\examples{
data(gene_ids)
head(gene_ids)
}
\keyword{datasets}
