% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpol_refdata.R
\name{get.spline}
\alias{get.spline}
\title{Interpolation of ICA component of reference time series}
\usage{
get.spline(ICA, time.series, comp, pred.x, span = 0.25, plot = T)
}
\arguments{
\item{ICA}{an ICA object (returned by icafast)}

\item{time.series}{the initial time series}

\item{comp}{the index of the ICA component on which to fit the curve}

\item{pred.x}{the new time series to predict component values on}

\item{span}{span argument given to loess}

\item{plot}{if TRUE, plot the component and its fitted curve}
}
\description{
function that fits a loess curve on an ICA component and returns predictions on given x series
}
