% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_compare.R
\name{get_refTP}
\alias{get_refTP}
\title{Get reference expression profiles matching sample age estimates.}
\usage{
get_refTP(ref, ae_obj = NULL, ae_values = NULL, return.idx = TRUE)
}
\arguments{
\item{ref}{a \code{ref} object (as returned by \link{\code{make_ref}})}

\item{ae_obj}{an \code{ae} object (as returned by \link{\code{ae}})}

\item{ae_values}{age estimate values (used if ae_obj is NULL).}

\item{return.idx}{if TRUE (default) returns reference indices. If FALSE, returns reference expression matrix}
}
\value{
either a vector of indices (if \code{return.idx} is True) or an expression matrix for selected reference time points.
}
\description{
Fetches the (indices of, or) expression profiles of a reference that correspond to the input age estimates.
Matching reference data can be used to quantify the developmental signal between
experimental groups where the variable of interest is confounded by development.
}
\examples{
\donttest{
requireNamespace('wormRef', quietly = TRUE)

# get sample gene expression data
X <- wormRef::Cel_larval$g[,1:9]

# get reference
ref <- prepare_refdata(ref = 'Cel_larval', datapkg = 'wormRef' , n.inter = 200)

# define groups
fac <- factor(c('a', 'a', 'b', 
                'a', 'b', 'b',
                'c', 'c', 'c'))

# estimate sample age
ae_X <- ae(X, ref)

# compare group diff. expr. with matching reference
rc <- ref_compare(X, ref, fac, ae_X)
print(rc)

# get sample and reference (ie. development) logFCs between groups
lfc_a_vs_b <- get_logFC(rc)
lfc_a_vs_c <- get_logFC(rc, l = 'c')
lfc_b_vs_c <- get_logFC(rc, l0='b', l = 'c')

# plot sample vs. reference logFCs
par(mfrow = c(2,2))
plot(ae_X, groups = fac)
plot(lfc_a_vs_b, main = 'a vs. b logFC')
plot(lfc_a_vs_c, main = 'a vs. c logFC')
plot(lfc_b_vs_c, main = 'b vs. c logFC')

}
}
