% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_formatting.R
\name{GPL200_to_WB}
\alias{GPL200_to_WB}
\title{Get a GPL200 dataset with WBGene ids}
\usage{
GPL200_to_WB(GEO_id = NULL, expr.matrix = NULL, raw = F)
}
\arguments{
\item{GEO_id}{a GEO accession number (e.g 'GSE52747'), must be of GPL200 platform format.}

\item{expr.matrix}{a raw gene expression matrix of of GPL200 platform format. Is ignored if a GEO accession number is given.}

\item{raw}{wether to return the raw data or perform log(expr.matrix +1) before output.}
}
\value{
an expression matrix with WBGene IDs as row names
}
\description{
This function either downloads a GPL200 GEO dataset from given accession number and transfers 
it to WBGene ids, or takes input Gene expression matrix \emph{of GPL200 format} to do the same thing.
**Warning :** Since not all Affymetrix array probes (GPL200 format) have a WBGene equivalent,
about 30% of the initial 22625 genes are lost.
}
\examples{
\donttest{ 
gorrepati_wb <- GPL200_to_WB(GEO_id="GSE52747")
head(gorrepati_wb)
}

}
