% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpol_refdata.R
\name{interpol_refdata}
\alias{interpol_refdata}
\title{Interpolation of gene expression on time series from reference data}
\usage{
interpol_refdata(
  X,
  n.inter,
  time.series,
  ica.nc,
  keep.c,
  center = T,
  t.min = NULL,
  t.max = NULL,
  new.timepoints = NULL,
  span = 0.25,
  plot = F,
  return.fits = F
)
}
\arguments{
\item{X}{gene expression matrix of reference time series, genes as rows, (ordered) individuals as columns}

\item{n.inter}{number of timepoints to return in interpolated data}

\item{time.series}{timepoints of the reference (X).}

\item{ica.nc}{number of components to keep in ica (see \code{\link[ica]{icafast}})}

\item{keep.c}{indices of components to keep for interpolation}

\item{center}{defaults to TRUE ; wether to perform a centered ICA (see \code{\link[ica]{icafast}}).}

\item{t.min, t.max}{defaults to min and max of \code{time.series} ; start and end times of new time series ; ignored if new.timepoints is given.}

\item{new.timepoints}{vector of length n.inter with the new time series (overrides t.min and t.max)}

\item{span}{value(s) given to loess for curve fitting on the ica components}

\item{plot}{if TRUE, plots the selected components and their fitted curves}

\item{return.fits}{if TRUE, returns the list of fitted loess objects and curves}
}
\description{
\strong{DEPRECATED please use plsr_interpol()} - function that computes interpolated data from a reference time series.
}
\examples{
# Deprecated.



}
