% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refs_functions.R
\name{list_refs}
\alias{list_refs}
\title{List the references in a data-package}
\usage{
list_refs(datapkg)
}
\arguments{
\item{datapkg}{the name of the data-package to list references from.}
}
\description{
This function looks for the \code{ref_list} object from the input data-package and prints it.
}
\examples{

# list_refs(datapkg = "wormRef")

}
