% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{log2tpmplusone_cutoffs_sens}
\alias{log2tpmplusone_cutoffs_sens}
\title{Cut-off values of predictor scores based on log2(TPM+1) gene expression data. Should not be used for expression data using other normalisation methods.}
\format{
A list of 4 numeric values.
\describe{
 \item{log2tpmplusone_cutoffs_sens}{Numeric values giving the cutoff to define high vs low probability of response groups}

}
}
\source{

}
\usage{
log2tpmplusone_cutoffs_sens
}
\description{
These cutoff values were derived from the maximum value of the specificity, with a sensitivity equal to 1 (so that no false negatives exist)
}
\keyword{datasets}
