% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mperf.R
\name{mperf}
\alias{mperf}
\title{Model performance}
\usage{
mperf(
  Y,
  Yh,
  global = TRUE,
  to_compute = c("aCC", "aRE", "MSE", "aRMSE"),
  is.t = FALSE
)
}
\arguments{
\item{Y, Yh}{The data and predictions matrices respectively with variables as columns, observations as rows ; must have the same dimensions.}

\item{global}{if TRUE (default), averages the index over all variables.}

\item{to_compute}{a vector with the indices to compute among c("aCC", "aRE", "MSE", "aRMSE").}

\item{is.t}{boolean ; if TRUE, Y and Yh should be transposed.}
}
\description{
Computes indices of model performance from real data and predictions.
}
\section{Indices}{


Let \eqn{y} and \eqn{\hat{y}} be the data and predictions respectively, with \eqn{m} dependant variables and \eqn{n} observations.
The model performance indices are defined as follows.

\describe{
\item{aCC}{average Correlation Coefficient.
\deqn{
  aCC = \frac{1}{m}\sum^{m}_{i=1}{CC} = \frac{1}{m}\sum^{m}_{i=1}{\cfrac{\sum^{n}_{j=1}{(y_i^{(j)}-\bar{y}_i)(\hat{y}_i^{(j)}-\bar{\hat{y}}_i)}}{\sqrt{\sum^{n}_{j=1}{(y_i^{(j)}-\bar{y}_i)^2(\hat{y}_i^{(j)}-\bar{\hat{y}}_i)^2}}}}
  }}

\item{aRE}{average Relative Error.
\deqn{
  a\delta = \frac{1}{m}\sum^{m}_{i=1}{\delta} = \frac{1}{m} \sum^{m}_{i=1} \frac{1}{n} \sum^{n}_{j=1} \cfrac{| y_i^{(j)} - \hat{y}_i^{(j)} | }{y_i^{(j)}}
  }}

\item{MSE}{Mean Squared Error.
\deqn{
  MSE = \frac{1}{m} \sum^{m}_{i=1} \frac{1}{n} \sum^{n}_{j=1} (y_i^{(j)} - \hat{y}_i^{(j)} )^2
  }}

\item{aRMSE}{average Root Mean Squared Error.
\deqn{
  aRMSE = \frac{1}{m}\sum^{m}_{i=1}{RMSE} = \frac{1}{m} \sum^{m}_{i=1} \sqrt{\cfrac{\sum^{n}_{j=1} (y_i^{(j)} - \hat{y}_i^{(j)} )^2}{n}}
  }}
}
}

\examples{

m1 <- matrix(rnorm(1000), ncol = 5)
m2 <- matrix(rnorm(1000), ncol = 5)
mperf(m1, m2, is.t = TRUE)

}
