% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oud_ref-data.R
\docType{data}
\name{oud_ref}
\alias{oud_ref}
\title{Oudenaarden reference data time series}
\format{a list with \code{X} being the gene expression matrix and \code{time.series} being the timepoints (in hours)}
\source{
\href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE49043}{GEO dataset}
}
\usage{
data(oud_ref)
}
\description{
Data from a paper published by Oudenaarden et al.
with C elegans gene expression levels (log+1 from rpkm) at 
26 time points.
}
\examples{
data(oud_ref)
times <- attr(oud_ref, "time.series")
gene1 <- oud_ref$X[,1]
\donttest{plot(times, gene1, type='l')}
}
\references{
Oudenaarden et al. (2013) Nat Genet. 2013 Nov; 45(11): 1337–1344. 
(\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3812263/}{PubMed})
}
\keyword{datasets}
