% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ae.R
\name{plot.ae}
\alias{plot.ae}
\title{Plot an ae object}
\usage{
\method{plot}{ae}(
  x,
  groups = NULL,
  subset = NULL,
  show.boot_estimates = T,
  show.prior = F,
  col = par("fg"),
  color = col,
  col.b = 2,
  col.p = 1,
  col.l = "gray",
  pch = 16,
  cex = 1,
  truncate_name = c("none", "end", "start"),
  sn_len = 10,
  lmar = 10,
  g.line = lmar * 0.75,
  xlim = NULL,
  xlab = NULL,
  main = NULL,
  CIbar.width = 0.1,
  glob.above = F,
  l.pos = "bottomright",
  ...
)
}
\arguments{
\item{x}{an \code{ae} object, as returned by \code{\link{ae}}.}

\item{groups}{a factor with sample categories (e.g. treatment groups).}

\item{subset}{an index vector of the samples to plot (defaults to all).}

\item{show.boot_estimates}{logical ; if TRUE (default), shows individual bootstrap estimates as swarms.}

\item{show.prior}{logical ; if TRUE, shows input prior(s) on the plot.}

\item{col, color}{the estimate, confidence interval, and label color.}

\item{col.b, col.p, col.l}{color of bootstrap estimates, priors, and background lines respectively.}

\item{pch, cex, xlim}{graphical parameters.}

\item{truncate_name}{whether to truncate displayed sample names from start, end, or not (default).}

\item{sn_len}{number of characters to keep when truncating sample names.}

\item{lmar}{left margin value, increase to fit sample names.}

\item{g.line}{position of the group names (margin line).}

\item{xlab}{x axis label.}

\item{main}{plot title.}

\item{CIbar.width}{the width of the confidence interval bars.}

\item{glob.above}{logical ; if TRUE, the global estimate is overlayed above all else.}

\item{l.pos}{position of the legend when show.prior is \code{TRUE}, passed on to \code{\link[graphics]{legend}}}

\item{...}{additional arguments passed on to \code{\link{points}}.}
}
\description{
Plots age estimates of samples with bootstrap confidence intervals.
}
\examples{
\donttest{
requireNamespace('wormRef', quietly = TRUE)

# get some samples to stage
samp <- wormRef::Cel_larval$g[,13:15]

# load interpolated reference
r_larv <- prepare_refdata(ref = 'Cel_larval', datapkg = 'wormRef' , n.inter = 200)

# perform age estimate
ae_test <- ae(samp, r_larv)

# check output
summary(ae_test)
plot(ae_test) # plot all sample estimates
plot_cor(ae_test) # plot individual correlation profiles of samples

# get results
ae_test$age.estimates
}
}
