% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_gene_expr.R
\name{plot.corg}
\alias{plot.corg}
\title{Plot a corg object}
\usage{
\method{plot}{corg}(cors, col = cm.colors(256), ...)
}
\arguments{
\item{cors}{a \code{corg} object, as returned by \code{\link{cor.gene_expr}}}

\item{col}{a color scheme, passed on to \code{\link{heatmap}}}

\item{...}{additional arguments passed on to \code{\link{heatmap}}}
}
\description{
Plots the correlation matrix returned by \code{\link{cor.gene_expr}} 
as a heatmap.
}
\examples{
data(oud_ref)

samp <- oud_ref$X[,c(3,8,12,20)]
cc <- cor.gene_expr(samp, oud_ref$X)
\donttest{
plot(cc, margins=c(10,5))
}

}
