% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dfCV.R
\name{plot.dfCV}
\alias{plot.dfCV}
\title{Plot a dfCV object}
\usage{
\method{plot}{dfCV}(
  x,
  stat = c("median", "mean"),
  subset = 1:length(x$dfs),
  t.plot = TRUE,
  t.col = makeTransparent("black", 30),
  signchange = TRUE,
  main = "CV Error",
  type = "b",
  cex = 0.5,
  lwd = 2,
  leg = TRUE,
  l.pos = "top",
  ...
)
}
\arguments{
\item{x}{a \code{dfCV} object, as returned by \code{\link{df_CV}}.}

\item{stat}{either "median" or "mean".}

\item{subset}{indices of a subset of df values to plot}

\item{t.plot}{boolean ; if TRUE, plots the individual CV Error trajectories.}

\item{t.col}{color for the individual CV Error trajectories.}

\item{signchange}{boolean ; if TRUE, displays '+'/'-' above the curve to indicate diff from previous point.}

\item{main}{title of the plot.}

\item{type}{the line type.}

\item{cex}{sizing parameter applied to various elements of the plot.}

\item{lwd}{the line width.}

\item{leg}{boolean ; if TRUE, displays a legend.}

\item{l.pos}{the position of the legend, as passed on to \code{\link[graphics]{legend}}}

\item{...}{additional arguments passed on to \code{\link{plot}}.}
}
\description{
Plots the dfCV object to help determine the optimal df parameter value for PLSR regression.
}
\examples{
\donttest{
data(Cel_embryo)

dfCVembryo <- df_CV(Cel_embryo$X, Cel_embryo$time.series, 
                    dfs = 3:17, cv.n = 50, cv.s = 0.8)
plot(dfCVembryo)
}


}
