% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ref_timelines.R
\name{plot_ref_timelines}
\alias{plot_ref_timelines}
\title{Plot the reference dataset time series}
\usage{
plot_ref_timelines()
}
\description{
Displays the reference datasets' coverage above key devlopmental stages.
The data used for plotting is in the \code{\link{ref_tables}} object.
}
\examples{
\donttest{
plot_ref_timelines()
}

}
