% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsr_interpol.R
\name{plsr_interpol}
\alias{plsr_interpol}
\title{Interpolation of gene expression on time series from reference data}
\usage{
plsr_interpol(X, time.series, df, covar = NULL, topred = NULL,
  n.inter = 200, tmin = min(time.series), tmax = max(time.series),
  scale = T, knots = NULL, return.model = FALSE)
}
\arguments{
\item{X}{gene expression matrix of reference time series, genes as rows, (ordered) individuals as columns.}

\item{time.series}{timepoints of the reference (X).}

\item{df}{the df parameter passed on to the \code{\link[splines]{ns}} function.}

\item{covar}{a covariate to include in the model (*e.g* batch).}

\item{topred}{a level of \code{covar} to use for model predictions ; defaults to the first level.}

\item{n.inter}{number of timepoints to return in interpolated data, defaults to 200.}

\item{tmin, tmax}{defaults to min and max of \code{time.series} ; start and end times of interpolated time series.}

\item{scale}{defaults to TRUE, passed on to the \code{\link[pls]{plsr}} function.}

\item{knots}{defaults to NULL, passed on to the \code{\link[splines]{ns}} function.}

\item{return.model}{if TRUE, returns the PLSR model object and the input df value.}
}
\description{
This function computes the interpolated gene expression data from a reference time series.
This is done with a multi-target Partial Least Square Regression (PLSR) model, using a spline of time as descriptive variables.
}
\examples{

data("Cel_embryo")

iGE <- plsr_interpol(Cel_embryo$X, Cel_embryo$time.series, df = 9, n.inter = 100)

\donttest{
# plot random genes with their interpolations
par(mfrow=c(2,2))
invisible(sapply(sample(1:nrow(Cel_embryo$X), 4), function(i){
  plot(Cel_embryo$time.series, Cel_embryo$X[i,], lwd = 2, 
       xlab = "time", ylab = "gene expression", 
       main = rownames(Cel_embryo$X)[i])
  points(iGE$time.series, iGE$interpGE[i,], lwd = 2,
         type = "l", col = "royalblue")
}))
}


}
