% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos.det.R
\name{pos.det}
\alias{pos.det}
\title{Detect position of tracheids in rows and radial files.}
\usage{
pos.det(input, swe = 0.5, sle = 3, ec = 1.75, swl = 0.25, sll = 5,
  lc = 5, prof.co = 6, max.cells = 0.5, yrs = FALSE, aligning = TRUE,
  make.plot = TRUE)
}
\arguments{
\item{input}{an input data frame as produced by first.cell().}

\item{swe}{numeric value that is multiplied by the square-rooted cell lumen
area (l) of the target cell and used to determine the width of the
rectangular search area which locates the next earlywood cell in the row (default = 0.5).}

\item{sle}{a numeric value that is multiplied by the square rooted cell lumen
area (l) of the target cell to determine the length of the rectangle search
area which locates the next earlywood cell in the row (default = 3).}

\item{ec}{threshold ratio between the lumen area of two consecutive earlywood
cells to determine the end of the earlywood search (default =
1.75). The default setting indicates that the earlywood search ends when
the next cell lumen area is at least 1.75 times smaller than the target
cell.}

\item{swl}{a numeric value that is multiplied by the square rooted cell lumen
area (l) of the target cell to determine the width of the rectangle search
area which locates the next latewood cell in the row (default = 0.25).}

\item{sll}{a numeric value that is multiplied by the square rooted cell lumen
area (l) of the target cell to determine the length of the rectangle search
area which locates the next earlywood cell in the row (default = 5).}

\item{lc}{threshold ratio between the lumen area of two consecutive latewood
cells to determine the end of the radial file. lc = 10 (default = 10) means
that the latewood search ends when the next cell lumen area is at least 10
times smaller than the target cell.}

\item{prof.co}{threshold ratio between the distance to the previous and
consecutive cell to determine if the row (or radial file) should be
excluded (default = 6).}

\item{max.cells}{threshold proportion of the maximum number of cells to
determine if the radial file has to be excluded (default = 0.6).}

\item{yrs}{either a vector providing the year(s) of interest or FALSE to
select all years included in input (Default = FALSE).}

\item{aligning}{logical. Indicates whether a second alignment has to be
performed based upon the cells detected within first.cell() (default =
TRUE).}

\item{make.plot}{logical. Indicating whether to make a plot (default =
FALSE).}

\item{list}{a data frame of CID (unique cell ID) from cells which should be
considered as first row cells. Should be filled if first.cell() output is
not used as input. The vector should be ordered to years and present
specific CID which is considered the first cell.}
}
\value{
Returns a original data frame with added columns for tracheid row and
  file position, as well as a marker column with diagnostic flags with
  additional.
}
\description{
Returns a data frame, i.e. the same used as for input and produced by
\code{first.cell()}, but with two added columns showing the \code{ROW} and
\code{POSITION} which indicate the radial files number (ROW, from left to
right) and radial cell position (from earlywood to latewood) of each tracheid.
NAs are assigned to tracheids not belonging to recognized radial files.
A \code{MARKER} column is added to indicate the last detected cell in the earlywood
search (indicated with 1), latewood search (2), the last detected cell (3),
rows that are removed due to gaps (4, see prof.co) and rows that are removed
due to limited amount of cells (5, see max.cells).
}
\examples{
#example of position detection
input<-is.raptor(example.data(species="MOUNT_PINUS") , str = FALSE)
aligned<-align(input,list=c("h","h","h",0.03),make.plot = FALSE)
first<-first.cell(aligned, frac.small = 0.2, yrs = FALSE, make.plot = FALSE)
output<-pos.det(first, swe = 0.7, sle = 3, ec = 1.75, swl = 0.5, sll = 5, lc = 10,prof.co = 1.7, max.cells = 0.7, yrs = FALSE , aligning = FALSE , make.plot = TRUE)
}
