% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ge_im.R
\name{print.geim}
\alias{print.geim}
\title{Print a geim object}
\usage{
\method{print}{geim}(x, ...)
}
\arguments{
\item{x}{a \code{geim} object, as returned by \code{\link{ge_im}}.}

\item{...}{arguments passed on to \code{\link{print}}}
}
\description{
Prints a \code{geim} object
}
