% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ge_imCV.R
\name{print.geimCV}
\alias{print.geimCV}
\title{Print a geimCV object}
\usage{
\method{print}{geimCV}(x, ...)
}
\arguments{
\item{x}{a \code{geimCV} object, as returned by \code{\link{ge_imCV}}.}

\item{...}{arguments passed on to \code{\link{print}}}
}
\description{
Prints a \code{geimCV} object
}
