% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_compare.R
\name{print.rcmp}
\alias{print.rcmp}
\title{Print an rcmp object}
\usage{
\method{print}{rcmp}(x, ...)
}
\arguments{
\item{x}{an \code{rcmp} object, as returned by \code{\link{ref_compare}}.}

\item{...}{arguments passed on to \code{\link{print}}}
}
\value{
invisibly returns a dataframe wiht the displayed info.
}
\description{
Prints an \code{rcmp} object
}
