% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ref.R
\name{print.ref}
\alias{print.ref}
\title{Print a ref object}
\usage{
\method{print}{ref}(x, ...)
}
\arguments{
\item{x}{a \code{ref} object, as returned by \code{\link{make_ref}}.}

\item{...}{arguments passed on to \code{\link{print}}}
}
\description{
Prints a \code{ref} object
}
