% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{proportions}
\alias{proportions}
\title{Proportions of response/non-response to stratify based on 4 predictors of immunotherapy benefit}
\format{
A list of 4 named numeric vectors.
\describe{
 \item{constant_list}{Numeric values giving the weight of each gene within each signature}

}
}
\source{

}
\usage{
proportions
}
\description{
The proportions of responders/NRs required to stratify based on the 4 probability of response scores.
}
\keyword{datasets}
