% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{proportions_sens}
\alias{proportions_sens}
\title{Proportions of response/non-response to stratify based on 4 predictors of immunotherapy benefit}
\format{
A list of 4 named numeric vectors.
\describe{
 \item{proportions_sens}{Numeric values giving the weight of each gene within each signature}

}
}
\source{

}
\usage{
proportions_sens
}
\description{
The proportions of responders/NRs required to stratify based on the 4 probability of response scores.
These proportions derive from cutoffs optimised by taking the threshold at the maximum value of specificity whilst sensitivity was equal to 1, thereby minimising false negatives.
}
\keyword{datasets}
