% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeRaptor.R
\name{raptorise}
\alias{raptorise}
\title{raptorise is a wrapper for the compute_risk_score function enabling it to be run over all 4 of the models}
\usage{
raptorise(expr)
}
\arguments{
\item{expr}{a data frame of gene expression data with genes as column names and samples as rownames.
Ideally expression will be in log2(TPM+1) format. The data frame should also include columns for the MSKCC prognostic score
if the data is from renal cancer patients (column name = "mskcc").

Note that the COX-IS model contains two coefficients that are 0. Ignore these. If they are missing they will be printed out as missing, but they aren't affecting the model anyway.}
}
\value{
data.frame with variables sample name and a column for each predictor method, with the continuous variable, the stratifier based on log2tpm expression data cutoff, and a stratifier based on the proportion of high/low patients in training.
}
\description{
raptorise is a wrapper for the compute_risk_score function enabling it to be run over all 4 of the models
}
