% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_table-data.R
\docType{data}
\name{ref_table}
\alias{ref_table}
\title{Reference Table}
\format{A dataframe with :
\itemize{
\item \code{name}: the reference name,
\item \code{organism}: the organism,
\item \code{data_pkg}: the data-package to load,
\item \code{description}: a brief description of the reference
}}
\usage{
ref_table
}
\description{
A dataframe listing the available pre-built references.
}
\keyword{datasets}
