% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sterken_ref-data.R
\docType{data}
\name{sterken_ref}
\alias{sterken_ref}
\title{Sterken reference data time series}
\format{a list with \code{X} being the gene expression matrix and \code{time.series} being the timepoints (in hours post-hatching)}
\source{
\href{https://www.ebi.ac.uk/arrayexpress/experiments/E-MTAB-7574/}{ArrayExpress dataset}
}
\usage{
data(sterken_ref)
}
\description{
Data from the ArrayExpress submitted by Mark G. Sterken in 2014, 
recently published by Van Sluijs et al. (2019).
Only samples of Standard N2 strain were kept here.
}
\details{
Time is in hours from hatching, scaled on the Oudenaarden 20C time series, 
re-estimated on the Reinke and Oudenaarden references.
}
\examples{
data(sterken_ref)
times <- sterken_ref$time.series
gene1 <- sterken_ref$X[700,]
\donttest{plot(times, gene1, type='b')}
}
\references{
Van Sluijs, et al.  (2019) BioRxiv : 579151
(\href{https://www.biorxiv.org/content/10.1101/579151v3}{bioRxiv})
}
\keyword{datasets}
