% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae.R
\name{summary.ae}
\alias{summary.ae}
\title{Print an ae object summary}
\usage{
\method{summary}{ae}(object, digits = 3, ...)
}
\arguments{
\item{object}{an \code{ae} object, as returned by \code{\link{ae}}.}

\item{digits}{the number of digits passed on to \code{\link{round}}}

\item{...}{ignored (needed to match the S3 standard)}
}
\value{
a list holding a data frame of ordered age estimates and confidence intervals, the span of estimates and the range.
}
\description{
Prints a summary of the \code{age.estimates} dataframe of an \code{ae} object
}
\examples{
\donttest{
requireNamespace('wormRef', quietly = TRUE)

# get some samples to stage
samp <- wormRef::Cel_larval$g[,13:15]

# load interpolated reference
r_larv <- prepare_refdata(ref = 'Cel_larval', datapkg = 'wormRef' , n.inter = 200)

# perform age estimate
ae_test <- ae(samp = samp, 
              refdata = r_larv$interpGE, 
              ref.time_series = r_larv$time.series)

# check output
summary(ae_test)
plot(ae_test, show.boot_estimates = TRUE) # plot all sample estimates
plot_cor.ae(ae_test) # plot individual correlation profiles of samples

# get results
ae_test$age.estimates
}
}
