% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_CV.R
\name{summary.dfCV}
\alias{summary.dfCV}
\title{dfCV object summary}
\usage{
\method{summary}{dfCV}(object, digits = 3, ...)
}
\arguments{
\item{object}{a \code{dfCV} object, as returned by \code{\link{df_CV}}.}

\item{digits}{the number of digits passed on to \code{\link{round}}}

\item{...}{ignored (needed to match the S3 standard)}
}
\value{
invisibly returns the string output
}
\description{
Prints a summary of the \code{dfCV} object
}
\examples{
\donttest{
data(Cel_embryo)

dfCVembryo <- df_CV(Cel_embryo$X, Cel_embryo$time.series, 
                    dfs = 3:17, cv.n = 50, cv.s = 0.8)
summary(dfCVembryo)
}

}
