% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_gene_expr.R
\name{cor.gene_expr}
\alias{cor.gene_expr}
\title{Compute correlation between sample and reference data}
\usage{
cor.gene_expr(samp, refdata, cor.method = "spearman")
}
\arguments{
\item{samp}{the sample matrix, gene as rows, individuals as columns}

\item{refdata}{the reference time series matrix, same format as \code{samp}}

\item{cor.method}{\code{method} parameter passed on to the \code{\link{cor}} function}
}
\value{
a matrix of correlation scores with samples as columns, refdata as rows
}
\description{
Computes the correlation between the gene expressions of the sample 
and the reference data using the \code{\link{cor}} function ;
default correlation index is spearman.
}
\examples{
data(oud_ref)

samp <- oud_ref$X[,c(3,8,12,20)]
cc <- cor.gene_expr(samp, oud_ref$X)
\donttest{
plot(cc, margins=c(10,5))
}

}
