% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified.R
\name{.stratified}
\alias{.stratified}
\title{Build a representative CV training set}
\usage{
.stratified(df, group, size, select = NULL, replace = FALSE, bothSets = FALSE)
}
\description{
All credit for this function goes to user mrdwab on GitHub :
\url{https://gist.github.com/mrdwab/6424112}
}
