% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_worm_age.R
\name{estimate.worm_age}
\alias{estimate.worm_age}
\title{Estimate the developmental age of individuals}
\usage{
estimate.worm_age(samp, refdata, ref.time_series,
  cor.method = "spearman", nb.cores = 2, bootstrap.n = 30,
  bootstrap.set_size = NULL, prior = NULL, prior.params = NULL,
  verbose = T)
}
\arguments{
\item{samp}{the sample matrix, gene as rows, individuals as columns}

\item{refdata}{the reference time series matrix, same format as \code{samp}}

\item{ref.time_series}{the reference time series (\emph{e.g.} \code{interpol$time.series} if using interpolated reference data)}

\item{cor.method}{correlation method argument passed on to \code{\link{cor.gene_expr}}. Note that the Spearman coefficient performs much better than Pearson (while a bit slower).}

\item{nb.cores}{the number of cores on which to parallelize the process, defaults to 2.}

\item{bootstrap.n}{the number of bootstrap steps. Should ideally be >5}

\item{bootstrap.set_size}{the size of the random sub genesets for the bootstrap, defaults to ngenes/3 (ngenes being the number of \emph{overlapping} genes between sample and reference).}

\item{prior}{a vector with an approximate development time of the samples, must be in the same units than \code{ref.time_series}. Vector is recycled if its length is smaller than the number of samples}

\item{prior.params}{the std. deviation of the prior scoring distribution. \emph{Note that setting this value too low can cause a significant bias in the age estimation.}}

\item{verbose}{boolean ; if TRUE, displays messages of the various steps of the method.}
}
\value{
an '\code{ae}' object, which is a list of the correlation matrix between sample and reference, the age estimates, the reference time series as well as the bootstrap correlation matrices and age estimates.
}
\description{
This function estimates the developmental age of individuals based on given
reference data as well as an approximate time estimate of the age.
This is achieved using a two-step process.
First, simple correlation between the gene expression profiles of
the sample and reference data are computed using \code{\link{cor.gene_expr}}; 
this gives correlation profiles according to reference time per individual. 
Then, the maxima of the correlation profiles are evaluated and scored according to 
a gaussian distribution around the time estimate given as input and their correlation score. 
The maxima scores range from 0 to 1, 1 being the case where the highest correlation peak is 
exactly at the given approximate time. 
The implemented bootstrap procedure re-estimates the age from the given times with a 
random uniform noise and returns the average best time, as well as the 95% quantile 
interval of the values.
}
\details{
Do note that using interpolated reference data (from \code{\link{interpol_refdata}})
gives the best results.
}
\examples{
data(oud_ref)

samp <- oud_ref$X[,13:15]
age.est <- estimate.worm_age(samp, oud_ref$X, oud_ref$time.series)
age.est$age.estimates
\donttest{
plot(age.est)
}


}
