% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_worm_age.R
\name{estimate.worm_age}
\alias{estimate.worm_age}
\title{Estimate the developmental age of individuals}
\usage{
estimate.worm_age(samp, refdata, ref.time_series, est.time, time.sd = 5,
  cor.method = "spearman", bootstrap.n = 50,
  bootstrap.time_window = 2, cors = NULL)
}
\arguments{
\item{samp}{the sample matrix, gene as rows, individuals as columns}

\item{refdata}{the reference time series matrix, same format as \code{samp}}

\item{ref.time_series}{the reference time series (\emph{e.g.} \code{interpol$time.series} if using interpolated reference data)}

\item{est.time}{a vector with the approximate development time of the samples, must be in the same units than \code{ref.time_series}. Vector is recycled if its length is smaller than the number of samples}

\item{time.sd}{the std. deviation of the gaussian scoring distribution. \emph{Note that setting this value too low can cause a significant bias in the age estimation.}}

\item{cor.method}{correlation method argument passed on to \code{\link{cor.gene_expr}}. Note that the spearman coefficient performs much better than the others.}

\item{bootstrap.n}{the number of re-estimates done by the bootstrap. If set to 0, the 95\% interval is computed from the reference time series' resolution}

\item{bootstrap.time_window}{the width of the window in which bootstrap re-estimates occur.}

\item{cors}{the correlation matrix between sample and reference data. \bold{This must be exactly} \code{cor.gene_expr(samp, refdata, cor.method = cor.method)}}
}
\value{
an '\code{ae}' object, which is a list of the correlation matrix between sample and reference, the age estimates, the initial time estimates and the reference time series.
}
\description{
This function estimates the developmental age of individuals based on given
reference data as well as an approximate time estimate of the age.
This is achieved using a two-step process.
First, simple correlation between the gene expression profiles of
the sample and reference data are computed using \code{\link{cor.gene_expr}}; 
this gives correlation profiles according to reference time per individual. 
Then, the maxima of the correlation profiles are evaluated and scored according to 
a gaussian distribution around the time estimate given as input and their correlation score. 
The maxima scores range from 0 to 1, 1 being the case where the highest correlation peak is 
exactly at the given approximate time. 
The implemented bootstrap procedure re-estimates the age from the given times with a 
random uniform noise and returns the average best time, as well as the 95% quantile 
interval of the values.
}
\details{
Do note that using interpolated reference data (from \code{\link{interpol_refdata}})
gives the best results.
}
\examples{
data(oud_ref)

samp <- oud_ref$X[,13:15]
age.est <- estimate.worm_age(samp, oud_ref$X, oud_ref$time.series, 26)
age.est$age.estimates
\donttest{
}


}
