% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_gene_expr.R
\name{estimate.worm_age}
\alias{estimate.worm_age}
\title{Estimate the developmental age of individuals}
\usage{
estimate.worm_age(samp, refdata, ref.time_series, est.time, time.sd = 5,
  cor.method = "pearson", all.peaks = F)
}
\arguments{
\item{samp}{the sample matrix, gene as rows, individuals as columns}

\item{refdata}{the reference time series matrix, same format as \code{samp}}

\item{ref.time_series}{the reference time series (\textit{e.g.} \code{colnames(refdata)} if using interpolated reference data)}

\item{est.time}{the approximate development time of the samples, must be in the same units than \code{ref.time_series}}

\item{time.sd}{the std. deviation of the gaussian scoring distribution. \textit{Note that setting this value too low can cause a significant bias in the age estimation.}}

\item{cor.method}{correlation method argument passed on to \code{\link{cor.gene_expr}}}

\item{all.peaks}{logical; if TRUE, returns all correlation peaks (potential age estimates) and their respective scores for every individuals, as a list. If FALSE, only returns the best estimate for each individual, as a dataframe.}
}
\value{
an '\code{ae}' object, which is a list of the correlation matrix between sample and reference, and the age estimates (either as list of individuals or dataframe, depending on \code{all.peaks}).
}
\description{
This function estimates the developmental age of individuals based on given
reference data as well as an approximate time estimate of the age.
This is achieved using a two-step process.
First, simple correlation between the gene expression profiles of
the sample and reference data are computed using \code{\link{cor.gene_expr}}; 
this gives correlation profiles according to reference time per individual. 
Then, the maxima of the correlation profiles are evaluated and scored according to 
a gaussian distribution around the time estimate given as input. 
The maxima scores range from 0 to 1, 1 being the case where a correlation peak is 
exactly at the given approximate time.
}
\examples{
data(oud_ref)

samp <- oud_ref$X[,13:15]
age.est <- estimate.worm_age(samp, oud_ref$X, oud_ref$time.series, 26)
age.est$age.estimates
\donttest{
}


}
