% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_worm_age.R
\name{estimate.worm_age}
\alias{estimate.worm_age}
\title{Estimate the developmental age of individuals}
\usage{
estimate.worm_age(samp, refdata, ref.time_series,
  cor.method = "spearman", nb.cores = 2, bootstrap.n = 30,
  bootstrap.set_size = NULL, prior = NULL, prior.params = NULL,
  verbose = T)
}
\arguments{
\item{samp}{the sample matrix, gene as rows, individuals as columns}

\item{refdata}{the reference time series matrix, same format as \code{samp}}

\item{ref.time_series}{the reference time series (\emph{e.g.} \code{interpol$time.series} if using interpolated reference data)}

\item{cor.method}{correlation method argument passed on to \code{\link{cor.gene_expr}}. Note that the Spearman coefficient performs much better than Pearson (while a bit slower).}

\item{nb.cores}{the number of cores on which to parallelize the process, defaults to 2.}

\item{bootstrap.n}{the number of bootstrap steps. Should ideally be >5}

\item{bootstrap.set_size}{the size of the random sub genesets for the bootstrap, defaults to ngenes/3 (ngenes being the number of \emph{overlapping} genes between sample and reference).}

\item{prior}{a vector with an approximate development time of the samples, must be in the same units than \code{ref.time_series}. Vector is recycled if its length is smaller than the number of samples}

\item{prior.params}{the std. deviation of the prior scoring distribution. \emph{Note that setting this value too low can cause a significant bias in the age estimation.}}

\item{verbose}{boolean ; if TRUE, displays messages of the various steps of the method.}
}
\value{
an '\code{ae}' object, which is a list of the correlation matrix between sample and reference, the age estimates, 
the reference time series as well as the bootstrap correlation matrices and age estimates.
There are plot, print and summary methods for this object.
}
\description{
This function estimates the developmental age of sample individuals based on 
correlation with given reference data. The correlation is computed through
the \code{\link{cor.gene_expr}} function.
}
\details{
A prior can be given to help with the estimate, in which case the peaks 
of the correlation profiles will be scored according to a gaussian
of the specified parameters.

The implemented bootstrap procedure re-estimates the age on random gene subsets
of fixed size to evaluate the robustness of the estimate, given in the form of
an interval holding 95 \% of the bootstrap age estimates. 
Since estimates can sometimes 'jump' from peak to peak during the bootstrap,
an \emph{imbalance} measure is computed on the intervals from the ratio of the 
distances from the global estimate (whole gene set) to the egdes of the interval.
If this ratio exceeds 5, it may be useful to give a prior for a more robust estimate.

Using interpolated reference data (from \code{\link{interpol_refdata}})
gives more precise results.
}
\examples{
data(oud_ref)

samp <- oud_ref$X[,13:15]
age.est <- estimate.worm_age(samp, oud_ref$X, oud_ref$time.series)
age.est$age.estimates
\donttest{
plot(age.est)
}

}
