% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_formatting.R
\name{format_ids}
\alias{format_ids}
\title{Format/convert the sample gene IDs}
\usage{
format_ids(X, IDs, from = 1, to = 2, aggr.fun = mean, verbose = TRUE)
}
\arguments{
\item{X}{the sample gene expression matrix with genes as rows and individuals as columns}

\item{IDs}{A dataframe holding current IDs for \code{X} and target IDs}

\item{from, to}{defaults to 1 and 2 ; columns of \code{IDs} holding current and target IDs of \code{X} respectively}

\item{aggr.fun}{the function used for \code{\link[stats]{aggregate}}}

\item{verbose}{if TRUE (default), prints number of genes kept and aggregated}
}
\value{
a gene expression matrix, with new IDs as rownames
}
\description{
This function converts the gene IDs from given input and aggregates results
if necessary, for resulting IDs to be unique.
}
\examples{
# make dummy data
ids <- data.frame(id.a = paste0('a', 1:10), id.b = paste0('b', 1:10), stringsAsFactors = FALSE)
X <- matrix(rnorm(50), ncol = 5)
rownames(X) <- ids$id.a


# format ids
format_ids(X, ids, from = "id.a", to = "id.b")

}
