% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ge_im.R
\name{ge_im}
\alias{ge_im}
\title{Gene expression interpolation model}
\usage{
ge_im(
  X,
  p,
  formula,
  method = c("gam", "glm", "limma"),
  dim_red = c("pca", "ica"),
  nc = ncol(X),
  ...
)
}
\arguments{
\item{X}{the gene expression matrix (genes as rows, samples as columns)}

\item{p}{a dataframe with the pheno data used in the formula (samples as rows) e.g. time, covariates.}

\item{formula}{the model formula, which must start with 'X ~'. See \code{\link[mgcv]{gam}}, \code{\link[stats]{glm}} or \code{\link[limma]{lmFit}} documentation for specifications.}

\item{method}{the model to fit, one of c("gam", "glm", "limma").}

\item{dim_red}{the dimension reduction method to use for interpolation, one of c("pca", "ica"), ignored when method is "limma".}

\item{nc}{the number of components to extract from \code{X} for interpolation, defaults to \code{ncol(X)}, ignored method is "limma".}

\item{...}{extra arguments passed on to model functions.}
}
\value{
a '\code{geim}' model object. This object has its predict method
}
\description{
Build a model to interpolate on a gene expression dataset.
This can be done either with gam or glm models fit on the components of a PCA or ICA.
It's also possible to have a limma model fit directly on the gene expression data.
}
