% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ge_imCV.R
\name{ge_imCV}
\alias{ge_imCV}
\title{Cross validation comparison of geim models}
\usage{
ge_imCV(
  X,
  p,
  formula_list,
  cv.n = 50,
  cv.s = 0.8,
  method = c("gam", "glm", "limma"),
  dim_red = c("pca", "ica"),
  nc = ncol(X),
  to_compute = c("aCC", "aRE", "MSE", "aRMSE"),
  nb.cores = 2,
  verbose = T,
  ...
)
}
\arguments{
\item{X}{the gene expression matrix (genes as rows, samples as columns)}

\item{p}{a dataframe with the pheno data used in the formula (samples as rows) e.g. time, covariates.}

\item{formula_list}{a list of model formulas to compare, which must start with 'X ~' (as passed on to \code{\link{ge_im}}).}

\item{cv.n}{number of cross-validation repeats.}

\item{cv.s}{ratio of samples to use for training set. If \code{cv.s > 1}, then \code{cv.s} samples are used for the training set.}

\item{method}{the model type to fit, one of c("gam", "glm", "limma").}

\item{dim_red}{the dimension reduction method to use for interpolation, one of c("pca", "ica"), ignored when method is "limma".}

\item{nc}{the number of components to extract from \code{X} for interpolation, defaults to \code{ncol(X)}, ignored method is "limma".}

\item{to_compute}{the model performance indices to compute during CV (see \code{\link{mperf}})}

\item{nb.cores}{the number of cores to use for parallel execution.}

\item{verbose}{boolean ; if TRUE, displays messages of the various steps of the method.}

\item{...}{extra arguments passed on to model functions.}
}
\description{
This function performs cross-validation (CV) with the aim of finding the optimal model from the given formulas.
Parameters are explored through the given list of formulas (\emph{e.g}, 'df=3' or 'df=4' must be specified in the formulas).
}
\details{
The CV training sets are defined to be representative of all variables included in the models.
This is done with a function attributed to GitHub user mrdwab \url{https://gist.github.com/mrdwab/6424112}.

Note that only one method/dimension reduction can be used at a time through this function.
}
