% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ref.R
\name{make_ref}
\alias{make_ref}
\title{Create a reference object}
\usage{
make_ref(
  m,
  from = NULL,
  to = NULL,
  cov.levels = NULL,
  n.inter = 500,
  by.inter = NULL,
  t.unit = "",
  metadata = list()
)
}
\arguments{
\item{m}{a geim model object (as returned by \code{\link{ge_im}}).}

\item{from, to}{start/end of the interpolation (defaults to first and last time points)}

\item{cov.levels}{a named list with potential model covariate levels (e.g batch, strain) to predict as (defaults to first level).}

\item{n.inter}{interpolation resolution, as in seq(start, end, length.out = n.inter). One of \code{n.inter} or \code{by.inter} must be specified.}

\item{by.inter}{interpolation resolution, as in seq(start, end, by = by.inter). One of \code{n.inter} or \code{by.inter} must be specified.}

\item{t.unit}{an optional string specifying the time unit and t-zero, e.g "h past egg-laying".}

\item{metadata}{an optional named list with reference metadata (e.g. organism, tissue).}
}
\value{
a '\code{ref}' object to use with ae() for age estimation.
}
\description{
Make a reference object from a geim model
}
\examples{
\donttest{
requireNamespace('wormRef', quietly = TRUE)
requireNamespace('stats', quietly = TRUE)

# gene expression data
X <- wormRef::Cel_larval$g

# pheno data (e.g time, batch)
p <- wormRef::Cel_larval$p

# do a pca & select nb of components to use for interpol
pca <- stats::prcomp(X, rank = 20)
nc <- sum(summary(pca)$importance[3, ] < .999) + 1


# find optimal spline type
# setup formulas
smooths <- c('bs', 'tp', 'cr', 'ds')
flist <- as.list(paste0('X ~ s(age, bs = \'', smooths, '\') + cov'))
# do CV
cvres <- ge_imCV(X = scale(X), p = p, formula_list = flist,
                 cv.n = 20, nc = nc)
# check results
plot(cvres, names.arrange = 4) # lowest pred error with 'ds' spline

# build model & interpolation data
m <- ge_im(X = X, p = p, formula = 'X ~ s(age, bs = \'ds\') + cov', nc = nc)
n.inter = 100
ndat <- data.frame(age = seq(min(p$age), max(p$age), l = n.inter),
                   cov = rep(p$cov[1], n.inter))

# check interpolation on pca components
pred_pca <- predict(m, ndat, as.c = T)

par(mfrow = c(2,2))
invisible(sapply(seq_len(nc), function(i){
  plot(p$age, pca$rotation[, i], xlab = 'age', ylab = 'PC', main = paste0('PC',i),
       lwd = (p$cov == p$cov[1]) + 1)
  points(ndat$age, pred_pca[, i], type = 'l', lwd = 2)
}))

# get interpolated GE matrix, as a reference
r_X <- list(interpGE = predict(m, ndat), time.series = ndat$age)

# test
ae_X <- ae(X, r_X$interpGE, r_X$time.series)
par(mfrow = c(1,1))
plot(p$age, ae_X$age.estimates[,1])


}
}
