% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_worm_age.R
\name{makeTransparent}
\alias{makeTransparent}
\title{Make a color transparent}
\usage{
makeTransparent(color, alpha = 100)
}
\arguments{
\item{color}{any color.}

\item{alpha}{the alpha channel value, (0:255) - from fully transparent to opaque}
}
\description{
Makes any given color(s) transparent
}
\examples{
\donttest{
plot(1:10, col=makeTransparent('firebrick', 120), pch=16, cex=2)
}

}
