% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ae.R
\name{plot.ae}
\alias{plot.ae}
\title{Plot an ae object}
\usage{
\method{plot}{ae}(
  x,
  errbar.width = 0.1,
  show.prior = F,
  col.p = 1,
  show.boot_estimates = F,
  col.b = 2,
  groups = NULL,
  subset = NULL,
  glob.above = F,
  color = par("fg"),
  pch = 16,
  cex = 1,
  xlim = NULL,
  xlab = "Estimated ages",
  l.pos = "bottomright",
  ...
)
}
\arguments{
\item{x}{an \code{ae} object, as returned by \code{\link{ae}}.}

\item{errbar.width}{the width of the error bars.}

\item{show.prior}{logical ; if TRUE, shows the input prior(s) on the plot.}

\item{col.p}{the color of the prior estimate marker.}

\item{show.boot_estimates}{logical ; if TRUE, shows the individual bootstrapped estimates on the plot as swarms.}

\item{col.b}{the color of the bootstrapped estimates.}

\item{groups}{a factor with sample categories, as passed on to \code{\link{dotchart}}.}

\item{subset}{an index vector of the samples to plot (defaults to all).}

\item{glob.above}{logical ; if TRUE, the global estimate is plotted above all else.}

\item{color}{color parameter passed on to \code{\link{dotchart}}.}

\item{pch}{the pch parameter passed on to \code{\link{dotchart}}.}

\item{cex}{sizing parameter applied to various elements of the plot.}

\item{xlim}{horizontal range for the plot, see \code{\link[graphics]{plot.window}}, for example}

\item{xlab}{the x axis label, passed on to \code{\link{dotchart}}.}

\item{l.pos}{the position of the legend when show.prior is \code{TRUE}, as passed on to \code{\link[graphics]{legend}}}

\item{...}{additional arguments passed on to \code{\link{dotchart}}.}
}
\description{
Plots the age estimates along with bootstrap error bars using a dotchart.
}
\examples{
\donttest{
requireNamespace('wormRef', quietly = TRUE)

# get some samples to stage
samp <- wormRef::Cel_larval$g[,13:15]

# load interpolated reference
r_larv <- prepare_refdata(ref = 'Cel_larval', datapkg = 'wormRef' , n.inter = 200)

# perform age estimate
ae_test <- ae(samp = samp, 
              refdata = r_larv$interpGE, 
              ref.time_series = r_larv$time.series)

# check output
summary(ae_test)
plot(ae_test, show.boot_estimates = TRUE) # plot all sample estimates
plot_cor.ae(ae_test) # plot individual correlation profiles of samples

# get results
ae_test$age.estimates
}
}
