% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_worm_age.R
\name{plot.ae}
\alias{plot.ae}
\title{Plot an ae object}
\usage{
\method{plot}{ae}(x, errbar.width = 0.1, show.prior = F, col.p = 1,
  show.boot_estimates = F, col.b = 2, groups = NULL, subset = NULL,
  pch = 16, cex = 1, xlim = NULL, xlab = "Estimated ages",
  l.pos = "bottomright", ...)
}
\arguments{
\item{x}{an \code{ae} object, as returned by \code{\link{estimate.worm_age}}.}

\item{errbar.width}{the width of the error bars.}

\item{show.prior}{logical ; if TRUE, shows the input prior(s) on the plot.}

\item{col.p}{the color of the prior estimate marker.}

\item{show.boot_estimates}{logical ; if TRUE, shows the individual bootstrapped estimates on the plot as swarms.}

\item{col.b}{the color of the bootstrapped estimates.}

\item{groups}{a factor with sample categories, as passed on to \code{\link{dotchart}}.}

\item{subset}{an index vector of the samples to plot (defaults to all).}

\item{pch}{the pch parameter passed on to \code{\link{dotchart}}.}

\item{cex}{sizing parameter applied to various elements of the plot.}

\item{xlim}{horizontal range for the plot, see \code{\link[graphics]{plot.window}}, for example}

\item{xlab}{the x axis label, passed on to \code{\link{dotchart}}.}

\item{l.pos}{the position of the legend when show.prior is \code{TRUE}, as passed on to \code{\link[graphics]{legend}}}

\item{...}{additional arguments passed on to \code{\link{dotchart}}.}
}
\description{
Plots the correlation score curves from samples against a reference series \code{\link{cor.gene_expr}}
}
\examples{
data(oud_ref)

samp <- oud_ref$X[,13:15]
age.est <- estimate.worm_age(samp, oud_ref$X, oud_ref$time.series)
\donttest{
plot(age.est)
}

}
