% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_worm_age.R
\name{plot.ae}
\alias{plot.ae}
\title{Plot an ae object}
\usage{
\method{plot}{ae}(age.est, subset = 1:ncol(age.est$cors),
  show.init_estimate = F, c.lwd = 2, bar.size = 2,
  mx.col = "firebrick", in.col = "royalblue", ot.col = "grey50", ...)
}
\arguments{
\item{age.est}{an \code{ae} object, as returned by \code{\link{estimate.worm_age}}}

\item{subset}{an index vector of the samples to plot (defaults to all)}

\item{show.init_estimate}{logical ; if TRUE, shows the initial time estimate on the plot}

\item{c.lwd}{line width for the correlation score curve}

\item{bar.size}{cex of the maxima bars}

\item{mx.col}{color of the best age estimate bar}

\item{in.col}{color of the initial estimate bar}

\item{ot.col}{color of the other maxima bars}

\item{...}{additional arguments passed on to \code{\link{plot}}}
}
\description{
Plots the correlation score curves from samples against a reference series \code{\link{cor.gene_expr}} 
as a heatmap.
}
\examples{
data(oud_ref)

samp <- oud_ref$X[,13:15]
age.est <- estimate.worm_age(samp, oud_ref$X, oud_ref$time.series, 26)
\donttest{
plot(age.est)
}

}
