% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_worm_age.R
\name{plot.ae}
\alias{plot.ae}
\title{Plot an ae object}
\usage{
\method{plot}{ae}(age_est, errbar.width = 0.1, show.init_estimate = T,
  col.i = 1, groups = NULL, pch = 16, cex = 1,
  xlab = "Estimated ages", ...)
}
\arguments{
\item{errbar.width}{the width of the error bars}

\item{show.init_estimate}{logical ; if TRUE, shows the initial time estimate(s) on the plot}

\item{col.i}{the color of the initial estimate marker.}

\item{groups}{a factor with sample categories, as passed on to \code{\link{dotchart}}.}

\item{pch}{the pch parameter passed on to \code{\link{dotchart}}.}

\item{cex}{sizing parameter applied to various elements of the plot.}

\item{...}{additional arguments passed on to \code{\link{dotchart}}}

\item{age.est}{an \code{ae} object, as returned by \code{\link{estimate.worm_age}}}
}
\description{
Plots the correlation score curves from samples against a reference series \code{\link{cor.gene_expr}}
}
\examples{
data(oud_ref)

samp <- oud_ref$X[,13:15]
age.est <- estimate.worm_age(samp, oud_ref$X, oud_ref$time.series, 26)
\donttest{
plot(age.est)
}

}
