% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ge_imCV.R
\name{plot.geimCV}
\alias{plot.geimCV}
\title{Plot geimCV results}
\usage{
\method{plot}{geimCV}(
  x,
  join.plots = TRUE,
  to_plot = colnames(x$cve[[1]]),
  names = as.character(x$formula_list),
  names.arrange = NULL,
  tcol = 1,
  swarm = T,
  swarmargs = list(pch = 16),
  main = NULL,
  ...
)
}
\arguments{
\item{x}{a geimCV object, as returned by \code{\link{ge_imCV}}}

\item{join.plots}{boolean ; if TRUE, sets up par(mfrow) appropriately}

\item{to_plot}{indices to plot, defaults to all computed indices.}

\item{names}{the labels of the models, defaults to model formulas.}

\item{names.arrange}{if defined, prints names with shifted height (like 'steps') by given number of labels}

\item{tcol}{text color for names (only when \code{names.arrange} is defined)}

\item{swarm}{boolean ; if TRUE, displays individual values on top of boxplots as swarms}

\item{swarmargs}{list of parameters given to \code{\link[beeswarm]{beeswarm}}}

\item{main}{title of the plots, pasted with indices on each plot.}

\item{...}{extra arguments passed on to \code{\link[graphics]{boxplot}}}
}
\value{
invisibly returns a list of boxplot objects that can be re-plotted through \code{\link[graphics]{bxp}}.
}
\description{
Plot geimCV results
}
