% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ge_imCV.R
\name{plot.geimCV}
\alias{plot.geimCV}
\title{Plot geimCV results}
\usage{
\method{plot}{geimCV}(
  x,
  join.plots = TRUE,
  to_plot = colnames(x$cve[[1]]),
  names = as.character(x$formula_list),
  names.arrange = NULL,
  tcol = 1,
  swarm = T,
  swarmargs = list(pch = 16),
  main = NULL,
  ...
)
}
\arguments{
\item{x}{a geimCV object, as returned by \code{\link{ge_imCV}}}

\item{join.plots}{boolean ; if TRUE, sets up par(mfrow) appropriately}

\item{to_plot}{indices to plot, defaults to all computed indices.}

\item{names}{the labels of the models, defaults to model formulas.}

\item{names.arrange}{if defined, prints names with shifted height (like 'steps') by given number of labels}

\item{tcol}{text color for names (only when \code{names.arrange} is defined)}

\item{swarm}{boolean ; if TRUE, displays individual values on top of boxplots as swarms}

\item{swarmargs}{list of parameters given to \code{\link[beeswarm]{beeswarm}}}

\item{main}{title of the plots, pasted with indices on each plot.}

\item{...}{extra arguments passed on to \code{\link[graphics]{boxplot}}}
}
\value{
invisibly returns a list of boxplot objects that can be re-plotted through \code{\link[graphics]{bxp}}.
}
\description{
Plot geimCV results
}
\examples{
\donttest{
requireNamespace('wormRef', quietly = TRUE)
requireNamespace('stats', quietly = TRUE)

# gene expression data
X <- wormRef::Cel_larval$g

# pheno data (e.g time, batch)
p <- wormRef::Cel_larval$p

# do a pca & select nb of components to use for interpol
pca <- stats::prcomp(X, rank = 20)
nc <- sum(summary(pca)$importance[3, ] < .999) + 1


# find optimal spline type
# setup formulas
smooths <- c('bs', 'tp', 'cr', 'ds')
flist <- as.list(paste0('X ~ s(age, bs = '', smooths, '') + cov'))
# do CV
cvres <- ge_imCV(X = scale(X), p = p, formula_list = flist,
                 cv.n = 20, nc = nc)
# check results
plot(cvres, names.arrange = 4) # lowest pred error with 'ds' spline

# build model & interpolation data
m <- ge_im(X = X, p = p, formula = 'X ~ s(age, bs = 'ds') + cov', nc = nc)
n.inter = 100
ndat <- data.frame(age = seq(min(p$age), max(p$age), l = n.inter),
                   cov = rep(p$cov[1], n.inter))

# check interpolation on pca components
pred_pca <- predict(m, ndat, as.c = T)

par(mfrow = c(2,2))
invisible(sapply(seq_len(nc), function(i){
  plot(p$age, pca$rotation[, i], xlab = 'age', ylab = 'PC', main = paste0('PC',i),
       lwd = (p$cov == p$cov[1]) + 1)
  points(ndat$age, pred_pca[, i], type = 'l', lwd = 2)
}))

# get interpolated GE matrix, as a reference
r_X <- list(interpGE = predict(m, ndat), time.series = ndat$age)

# test
ae_X <- ae(X, r_X$interpGE, r_X$time.series)
par(mfrow = c(1,1))
plot(p$age, ae_X$age.estimates[,1])


}
}
