% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ae.R
\name{plot_cor.ae}
\alias{plot_cor.ae}
\title{(DEPRECATED) Plot an ae object}
\usage{
plot_cor.ae(
  age.est,
  subset = 1:ncol(age.est$cors),
  show.prior = F,
  c.lwd = 2,
  bar.size = 1,
  mx.col = "firebrick",
  col.p = "royalblue",
  xlab = NULL,
  ...
)
}
\arguments{
\item{age.est}{an \code{ae} object, as returned by \code{\link{ae}}}

\item{subset}{an index vector of the samples to plot (defaults to all)}

\item{show.prior}{logical ; if TRUE, shows the input prior(s) on the plot.}

\item{c.lwd}{line width for the correlation score curve}

\item{bar.size}{size of the estimate confidence interval bars}

\item{mx.col}{color of the age estimate bars}

\item{col.p}{color of the prior bar}

\item{xlab}{the x axis label, passed on to \code{\link{plot}}}

\item{...}{additional arguments passed on to \code{\link{plot}}}
}
\description{
This function has been renamed. Please use \code{\link{plot_cor}} instead.
}
\examples{
\donttest{
requireNamespace('wormRef', quietly = TRUE)

# get some samples to stage
samp <- wormRef::Cel_larval$g[,13:15]

# load interpolated reference
r_larv <- prepare_refdata(ref = 'Cel_larval', datapkg = 'wormRef' , n.inter = 200)

# perform age estimate
ae_test <- ae(samp, r_larv)

# check output
summary(ae_test)
plot(ae_test) # plot all sample estimates
plot_cor(ae_test) # plot individual correlation profiles of samples

# get results
ae_test$age.estimates
}
}
