% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_worm_age.R
\name{plot_cor.ae}
\alias{plot_cor.ae}
\title{Plot an ae object}
\usage{
plot_cor.ae(age.est, subset = 1:ncol(age.est$cors), show.prior = F,
  c.lwd = 2, bar.size = 1, mx.col = "firebrick",
  col.p = "royalblue", ...)
}
\arguments{
\item{age.est}{an \code{ae} object, as returned by \code{\link{estimate.worm_age}}}

\item{subset}{an index vector of the samples to plot (defaults to all)}

\item{show.prior}{logical ; if TRUE, shows the input prior(s) on the plot.}

\item{c.lwd}{line width for the correlation score curve}

\item{bar.size}{size of the estimate 95IC bars}

\item{mx.col}{color of the age estimate bars}

\item{col.p}{color of the prior bar}

\item{...}{additional arguments passed on to \code{\link{plot}}}
}
\description{
Plots the correlation score curves from samples against the reference series.
}
\examples{
data(oud_ref)

samp <- oud_ref$X[,13:15]
\donttest{
age.est <- estimate.worm_age(samp, oud_ref$X, oud_ref$time.series)
plot_cor.ae(age.est)
}

}
