% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_refs.R
\name{plot_refs}
\alias{plot_refs}
\title{Plot reference timelines}
\usage{
plot_refs(organism)
}
\arguments{
\item{organism}{the organism name; can be abbreviated.}
}
\description{
Plots a timeline chart for references of the chosen organism
}
\examples{
\donttest{
plot_refs("Caenorhabditis")
}

}
