% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_refdata.R
\name{prepare_refdata}
\alias{prepare_refdata}
\title{Prepare the reference data included in the package}
\usage{
prepare_refdata(ref, datapkg, n.inter = 200)
}
\arguments{
\item{ref}{the name of the reference dataset to load (as given in \code{\link{list_refs}}).}

\item{datapkg}{the name of the data-package to load the rreference from.}

\item{n.inter}{the resolution of the interpolation, as in \code{seq(start, end, length.out = n.inter)}.}
}
\value{
a list with the interpolated reference dataset and its associated time series.
}
\description{
This function loads (and interpolates on) pre-built reference datasets.
The available datasets of a valid data-package can be listed with \code{\link{list_refs}}.
}
\examples{
\donttest{
requireNamespace('wormRef', quietly = TRUE)

# get some samples to stage
samp <- wormRef::Cel_larval$g[,13:15]

# load interpolated reference
r_larv <- prepare_refdata('Cel_larval', n.inter = 200)

# perform age estimate
ae_test <- ae(samp = samp, 
              refdata = r_larv$interpGE, 
              ref.time_series = r_larv$time.series)

# check output
summary(ae_test)
plot(ae_test, show.boot_estimates = TRUE) # plot all sample estimates
plot_cor.ae(ae_test) # plot individual correlation profiles of samples

# get results
ae_test$age.estimates
}
}
\seealso{
\link{list_refs} \link{ge_im}
}
