% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_refdata.R
\name{prepare_refdata}
\alias{prepare_refdata}
\title{Prepare the reference data included in the package}
\usage{
prepare_refdata(ref = c("young_adult", "Cel_YA_adult1", "sterken",
  "larval_development", "Cel_larval", "oudenaarden",
  "embryonic_development", "Cel_embryo", "hashimshony", "Cel_YA_adult2",
  "reinke"), n.inter = 200)
}
\arguments{
\item{ref}{which reference dataset to load. Can be abbreviated.}

\item{n.inter}{the \code{n.inter} parameter, as passed on to \code{\link{interpol_refdata}}.}
}
\value{
the interpolated reference dataset, as would \code{\link{interpol_refdata}}
}
\description{
This function loads the desired reference dataset from the package and
performs the interpolation with optimal parameters.
Four available reference datasets are available
(embryonic development, larval development and young adult/adult - sterken or reinke).
}
\examples{
\donttest{ 
interpol.larval <- prepare_refdata(ref="larval")
}

}
\seealso{
[interpol_refdata()]
}
