% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae.R
\name{print.ae}
\alias{print.ae}
\title{Print an ae object}
\usage{
\method{print}{ae}(x, digits = 3, ...)
}
\arguments{
\item{x}{an \code{ae} object, as returned by \code{\link{estimate.worm_age}}.}

\item{digits}{the number of digits passed on to \code{\link{round}}}

\item{...}{arguments passed on to \code{\link{print}}}
}
\description{
Prints the \code{age.estimates} dataframe of an \code{ae} object
}
\examples{
\donttest{
requireNamespace('wormRef', quietly = TRUE)

# get some samples to stage
samp <- wormRef::Cel_larval$g[,13:15]

# load interpolated reference
r_larv <- prepare_refdata(ref = 'Cel_larval', datapkg = 'wormRef' , n.inter = 200)

# perform age estimate
ae_test <- ae(samp = samp, 
              refdata = r_larv$interpGE, 
              ref.time_series = r_larv$time.series)

# check output
summary(ae_test)
plot(ae_test, show.boot_estimates = TRUE) # plot all sample estimates
plot_cor.ae(ae_test) # plot individual correlation profiles of samples

# get results
ae_test$age.estimates
}
}
