% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_worm_age.R
\name{print.ae}
\alias{print.ae}
\title{Print an ae object}
\usage{
\method{print}{ae}(x, digits = 3, ...)
}
\arguments{
\item{x}{an \code{ae} object, as returned by \code{\link{estimate.worm_age}}.}

\item{digits}{the number of digits passed on to \code{\link{round}}}
}
\description{
Prints the \code{age.estimates} dataframe of an \code{ae} object
}
\examples{
data(oud_ref)

samp <- oud_ref$X[,13:15]
age.est <- estimate.worm_age(samp, oud_ref$X, oud_ref$time.series)

print(age.est)

}
