% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_CV.R
\name{print.dfCV}
\alias{print.dfCV}
\title{dfCV object print}
\usage{
\method{print}{dfCV}(x, digits = 3, ...)
}
\arguments{
\item{x}{a \code{dfCV} object, as returned by \code{\link{df_CV}}.}

\item{digits}{the number of digits passed on to \code{\link{round}}}

\item{...}{ignored (needed to match the S3 standard)}
}
\value{
invisibly returns the string output
}
\description{
Prints a summary of the \code{dfCV} object  (same as summary)
}
\examples{
\donttest{
requireNamespace('wormRef', quietly = TRUE)

# gene expression data
G <- wormRef::Cel_larval$g

# pheno data (e.g time, batch)
P <- wormRef::Cel_larval$p



# find optimal spline degree of freedom for PLSR interpolation
dfCV_G <- df_CV(X = G, 
                time.series = P$age,
                covar = P$cov,
                dfs = 10:20)

# check minimum of CV error
summary(dfCV_G)
plot(dfCV_G) # min for df = 17

# perform PLSR interpolation using optimal parameter
r_G <- plsr_interpol(X = G, 
                     time.series = P$age,
                     covar = P$cov,
                     df = 17) 

}
}
