% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_tables-data.R
\docType{data}
\name{ref_tables}
\alias{ref_tables}
\title{Reference time tables}
\format{a list of data frames with the developmental stages and times.}
\source{
The timings for developmental stages are taken from the datasets' papers.
}
\usage{
data(ref_tables)
}
\description{
A list of reference tables giving the approximate timings of key developmental 
stages in the reference datasets, as well as the dataset spans.
}
\examples{
\donttest{
plot_timelines()
}
}
\keyword{datasets}
