% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae.R
\name{summary.ae}
\alias{summary.ae}
\title{Print an ae object summary}
\usage{
\method{summary}{ae}(object, digits = 3, ...)
}
\arguments{
\item{object}{an \code{ae} object, as returned by \code{\link{estimate.worm_age}}.}

\item{digits}{the number of digits passed on to \code{\link{round}}}

\item{...}{ignored (needed to match the S3 standard)}
}
\value{
a list holding a data frame of ordered age estimates and confidence intervals, the span of estimates and the range.
}
\description{
Prints a summary of the \code{age.estimates} dataframe of an \code{ae} object
}
\examples{
\donttest{
data(Cel_larval)

samp <- Cel_larval$X[,13:15]
age.est <- estimate.worm_age(samp, Cel_larval$X, Cel_larval$time.series)

summary(age.est)
}

}
